총급여 실수령액 등 완료

+++ src/main/java/com/in4people/bootrestapi/salary/dto/MonthlyWageDTO.java
@@ -15,16 +15,16 @@ public class MonthlyWageDTO {
     // 사원번호 MEM_CODE memCode VARCHAR2(100) 사원테이블에서~ FK
 //    private String memCode;
     private int monthlyPaycheck;                         // 월급
-    private int netSalary;                                  // 실수령액
+    private double netSalary;                                  // 실수령액
     private Date paymentDate;                               // 지급일
     private int bonusSUM;                                   // 상여금합계
 
     private int totalSum;
-    private int nationalPension;
-    private int healthInsurance;
-    private int employmentInsurance;
-    private int incomeTax1;
-    private int incomeTax2;
+    private double nationalPension;
+    private double healthInsurance;
+    private double employmentInsurance;
+    private double incomeTax1;
+    private double incomeTax2;
 
     private String salarySettingCode;
